refactor(helpers): mactrick function to helper DRY
