+++ src/eletor/create_hs.py
@@ -2,147 +2,47 @@ from scipy.special import kv
 import numpy as np
 import math
 
+from eletor.helper import mactrick
+
 try:
     from numba import njit
 except ModuleNotFoundError:
     njit = lambda x: x
 
-@njit
-def mactrick(m,gamma_x):
-    """ Use Cholesky decomposition to get first column of C
-
-    Args:
-        m (int) : length of time series
-        gamma_x (float array) : first column of covariance matrix
-
-    Returns:
-        h : array of float with impulse response
-    """
-
-    U = np.zeros((m,2))  # C = U’*U
-    V = np.zeros((m,2))
-    h = np.zeros(m)
-
-    #--- define the generators u and v
-    U[:,0] = gamma_x/math.sqrt(gamma_x[0])
-    V[1:m,0] = U[1:m,0]
-    h[m-1] = U[m-1,0]
-
-    k_old =0;
-    k_new =1;
-    for k in range(0,m-1):
-        sin_theta = V[k+1,k_old]/U[k,k_old]
-        cos_theta = math.sqrt(1.0-pow(sin_theta,2))
-        U[k+1:m,k_new] = ( U[k:m-1,k_old] - sin_theta*V[k+1:m,k_old])/cos_theta
-        V[k+1:m,k_new] = (-sin_theta*U[k:m-1,k_old] + V[k+1:m,k_old])/cos_theta
-        h[m-1-k] = U[m-1,k_new]
-
-        k_old = 1-k_old
-        k_new = 1-k_new
-
-    return h
-
-def preprocess_params(plist, mname):
-    def preprocessor(*args, **kwargs):
-        args = list(args)
-        try:
-            for i in plist:
-                if i not in kwargs: kwargs[i] = args.pop(0)
-        except (IndexError, KeyError):
-            raise TypeError(f"{mname} noise model requires arguments {plist}")
-        return kwargs
-    return preprocessor
-
-_paramlist = {'White':('m','sigma'),
-              'Powerlaw':('m','sigma','kappa','units','dt'),
-              'Flicker':('m','sigma','units','dt'),
-              'RandomWalk':('m','sigma','units','dt'),
-              'GGM':('m','sigma','kappa','one_minus_phi','units','dt'),
-              'VaryingAnnual':('m','sigma','phi','units','dt'),
-              'Matern':('m','sigma','lamba','kappa'),
-              'AR1':('m','sigma','phi')}
-
-def create_h(noisemodel,*args,**kwargs):
-    """ Create impulse function
-    GONZA: Esto es un dispatcher que se encarga de llamar a la funcion de crea ion de ruido correspondiente el primer argumento es el modelo y los subsiguientes son los parametros del modelo
-
-    Args:
-        noisemodel : model name, next arguments depends on this parameter:
-
-        Model dependent parameters:
-            White':('m','sigma'),
-            Powerlaw':('m','sigma','kappa','units','dt'),
-            Flicker':('m','sigma','units','dt'),
-            RandomWalk':('m','sigma','units','dt'),
-            GGM':('m','sigma','kappa','one_minus_phi','units','dt'),
-            VaryingAnnual':('m','sigma','phi','units','dt'),
-            Matern':('m','sigma','lamba','kappa'),
-            AR1':('m','sigma','phi')
-
-        m (int) : length of time series
-        sigma (float) : noise variance
-        kappa (float) : spectral index
-        one_minus_phi (float) : (1-phi) for gauss markogv models
-        phi (float) : phi for periodic models
-        lamba (float) : lambda for mattern noise
-        units (string) : 'mom' or 'msf' (for years or hours)
-        dt (float) : sampling period in days
-    Returns:
-        sigma, h : noise amplitude + array of float with impulse response
-    """
-        
-    ## GONZA: est funcion se encarga de verificar si los parmteros que el modelo requiere estan en la lista de paramatros que se le pasa a la funcion
-    ## TODO: esta responsabilidad deberia estar en la funcion despachada
-    #try:
-    #    preprocessor = preprocess_params(_paramlist[noisemodel], noisemodel)
-    #except KeyError:
-    #    raise ValueError('Unknown noisemodel: {0:s}'.format(noisemodel))
-
-    #kwargs = preprocessor(*args,**kwargs)
-
-    #if noisemodel=='White':
-    #    return kwargs.get('sigma'), \
-    #            create_h_white(**kwargs)
-
-    #elif noisemodel == 'Powerlaw':
-    #    kwargs['spectral_density'] = -kwargs['kappa']/2.0
-
-    #    return gauss_markov_scale_variance(**kwargs),\
-    #           create_h_Powerlaw(**kwargs)
-
-    #elif noisemodel == 'Flicker':
-    #    kwargs['spectral_density'] = 0.5
-
-    #    return gauss_markov_scale_variance(**kwargs),\
-    #           create_h_Flicker(**kwargs)
-
-    #elif noisemodel == 'RandomWalk':
-    #    kwargs['spectral_density'] = 1.0
-
-    #    return gauss_markov_scale_variance(**kwargs),\
-    #           create_h_RandomWalk(**kwargs)
-
-    #elif noisemodel == 'GGM':
-    #    return create_h_GGM(**kwargs)
-    #    #kwargs['spectral_density'] = -kwargs['kappa']/2.0
-
-    #    #return gauss_markov_scale_variance(**kwargs),\
-    #    #       create_h_GGM(**kwargs)
-
-    #elif noisemodel=='VaryingAnnual':
-    #    return kwargs.get('sigma'), \
-    #           create_h_VaryingAnnual(**kwargs)
-
-    #elif noisemodel=='Matern':
-    #    return kwargs.get('sigma'), \
-    #           create_h_Matern(**kwargs)
-
-
-    #elif noisemodel=='AR1':
-    #    return kwargs.get('sigma'), \
-    #           create_h_AR1(**kwargs)
-
-    assert False # Shouldn't get here
+#@njit
+#def mactrick(m,gamma_x):
+#    """ Use Cholesky decomposition to get first column of C
+#
+#    Args:
+#        m (int) : length of time series
+#        gamma_x (float array) : first column of covariance matrix
+#
+#    Returns:
+#        h : array of float with impulse response
+#    """
+#
+#    U = np.zeros((m,2))  # C = U’*U
+#    V = np.zeros((m,2))
+#    h = np.zeros(m)
+#
+#    #--- define the generators u and v
+#    U[:,0] = gamma_x/math.sqrt(gamma_x[0])
+#    V[1:m,0] = U[1:m,0]
+#    h[m-1] = U[m-1,0]
+#
+#    k_old =0;
+#    k_new =1;
+#    for k in range(0,m-1):
+#        sin_theta = V[k+1,k_old]/U[k,k_old]
+#        cos_theta = math.sqrt(1.0-pow(sin_theta,2))
+#        U[k+1:m,k_new] = ( U[k:m-1,k_old] - sin_theta*V[k+1:m,k_old])/cos_theta
+#        V[k+1:m,k_new] = (-sin_theta*U[k:m-1,k_old] + V[k+1:m,k_old])/cos_theta
+#        h[m-1-k] = U[m-1,k_new]
+#
+#        k_old = 1-k_old
+#        k_new = 1-k_new
+#
+#    return h
 
 def create_h_white(
         *,
